Moved rendertargets into their own subfolders
