+++ gameengine/gameengine.vcxproj
@@ -207,12 +207,13 @@
     <ClInclude Include="src\render\renderer\RendererType.h" />
     <ClInclude Include="src\render\renderer\WrapperRenderer.h" />
     <ClInclude Include="src\render\RenderMode.h" />
-    <ClInclude Include="src\render\rendertarget\EngineCanvas.h" />
-    <ClInclude Include="src\render\rendertarget\RenderTarget.h" />
-    <ClInclude Include="src\render\rendertarget\RenderTargetConfig.h" />
-    <ClInclude Include="src\render\rendertarget\RenderTargetMode.h" />
-    <ClInclude Include="src\render\rendertarget\RenderTexture.h" />
-    <ClInclude Include="src\render\rendertarget\RenderTextureData.h" />
+    <ClInclude Include="src\render\rendertarget\canvas\EngineCanvas.h" />
+    <ClInclude Include="src\render\rendertarget\target\RenderTarget.h" />
+    <ClInclude Include="src\render\rendertarget\target\RenderTargetConfig.h" />
+    <ClInclude Include="src\render\rendertarget\target\RenderTargetMode.h" />
+    <ClInclude Include="src\render\rendertarget\texture\RenderTexture.h" />
+    <ClInclude Include="src\render\rendertarget\texture\RenderTextureGroup.h" />
+    <ClInclude Include="src\render\rendertarget\texture\RenderTextureGroupFilter.h" />
     <ClInclude Include="src\render\ShaderProgram.h" />
     <ClInclude Include="src\render\TargetType.h" />
     <ClInclude Include="src\render\texture\Texture.h" />
@@ -271,12 +272,12 @@
     <ClCompile Include="src\render\renderer\Renderer.cpp" />
     <ClCompile Include="src\render\renderer\RendererType.cpp" />
     <ClCompile Include="src\render\renderer\WrapperRenderer.cpp" />
-    <ClCompile Include="src\render\rendertarget\EngineCanvas.cpp" />
-    <ClCompile Include="src\render\rendertarget\RenderTarget.cpp" />
-    <ClCompile Include="src\render\rendertarget\RenderTargetConfig.cpp" />
-    <ClCompile Include="src\render\rendertarget\RenderTargetMode.cpp" />
-    <ClCompile Include="src\render\rendertarget\RenderTexture.cpp" />
-    <ClCompile Include="src\render\rendertarget\RenderTextureData.cpp" />
+    <ClCompile Include="src\render\rendertarget\canvas\EngineCanvas.cpp" />
+    <ClCompile Include="src\render\rendertarget\target\RenderTarget.cpp" />
+    <ClCompile Include="src\render\rendertarget\target\RenderTargetConfig.cpp" />
+    <ClCompile Include="src\render\rendertarget\target\RenderTargetMode.cpp" />
+    <ClCompile Include="src\render\rendertarget\texture\RenderTexture.cpp" />
+    <ClCompile Include="src\render\rendertarget\texture\RenderTextureGroup.cpp" />
     <ClCompile Include="src\render\ShaderProgram.cpp" />
     <ClCompile Include="src\render\TargetType.cpp" />
     <ClCompile Include="src\render\texture\Texture.cpp" />
