+++ gameengine/src/render/controllers/CubemapController.cpp
@@ -4,7 +4,7 @@
 #include <stdexcept>
 
 #include "../../Engine.h"
-#include "../rendertarget/RenderTexture.h"
+#include "../rendertarget/texture/RenderTexture.h"
 #include "../../scene/Scene.h"
 #include "../../scene/Camera.h"
 #include "../renderer/Renderer.h"
