+++ gameengine/src/render/controllers/ShadowController.cpp
@@ -1,11 +1,11 @@
 #include "ShadowController.h"
 
-#include "../rendertarget/RenderTexture.h"
+#include "../rendertarget/texture/RenderTexture.h"
 #include "../../scene/model/Reflection.h"
 #include "../../scene/Cubemap.h"
 #include "../renderer/WrapperRenderer.h"
 #include "../TargetType.h"
-#include "../rendertarget/RenderTextureData.h"
+#include "../rendertarget/texture/RenderTextureGroup.h"
 
 std::unique_ptr<Renderer> ShadowController::GenerateRenderer(int layer)
 {
