+++ gameengine/src/render/controllers/SkyboxController.cpp
@@ -1,11 +1,11 @@
 #include "SkyboxController.h"
 
-#include "../rendertarget/RenderTexture.h"
+#include "../rendertarget/texture/RenderTexture.h"
 #include "../../scene/model/Reflection.h"
 #include "../../scene/Cubemap.h"
 #include "../renderer/NormalRenderer.h"
 #include "../TargetType.h"
-#include "../rendertarget/RenderTextureData.h"
+#include "../rendertarget/texture/RenderTextureGroup.h"
 
 std::unique_ptr<Renderer> SkyboxController::GenerateRenderer(int layer)
 {
