+++ gameengine/src/render/rendertarget/texture/RenderTextureGroupFilter.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <vector>
+
+struct RenderTextureGroupFilter
+{
+	std::vector<int> colour;
+	bool depth = false;
+};
